- Refactoring the project architecture - Enhancing Dependency Injection - Changing namings - WIP: 100% test coverage
