--- stock-service/cmd/api/routes_test.go
@@ -1,35 +1,0 @@
-package main
-
-import (
-	"github.com/go-chi/chi/v5"
-	"net/http"
-	"testing"
-)
-
-func Test_routes_exist(t *testing.T) {
-	testApp := Config{}
-
-	testRoutes := testApp.routes()
-	chiRoutes := testRoutes.(chi.Router)
-
-	routes := []string{"/stock-product", "/stock-product/{productCode}"}
-
-	for _, route := range routes {
-		routeExists(t, chiRoutes, route)
-	}
-}
-
-func routeExists(t *testing.T, routes chi.Router, route string) {
-	found := false
-
-	_ = chi.Walk(routes, func(method string, foundRoute string, handler http.Handler, middlewares ...func(handler2 http.Handler) http.Handler) error {
-		if route == foundRoute {
-			found = true
-		}
-		return nil
-	})
-
-	if !found {
-		t.Errorf("did not found %s in registered routes", route)
-	}
-}
