+++ stock-service/model/ProductStock.go
@@ -1,0 +1,7 @@
+package model
+
+type ProductStock struct {
+	ProductCode string `json:"productCode"`
+	ProductName string `json:"productName"`
+	Quantity    int    `json:"quantity"`
+}
