+++ stock-service/service/product_stock_service_test.go
@@ -1,0 +1,13 @@
+package service
+
+import (
+	"testing"
+)
+
+func TestProductStockService_StoreNewProductStock(t *testing.T) {
+
+}
+
+func TestProductStockService_DeleteProductStock(t *testing.T) {
+
+}
